Condense types
