--- src/Die/guards.ts
@@ -1,4 +1,0 @@
-import { CustomSides } from '../types/options'
-
-export const isCustomSides = (sides: unknown): sides is CustomSides =>
-  Array.isArray(sides)
