+++ src/Die/single-die.ts
@@ -1,6 +1,3 @@
-import { CustomSides } from '../types/options'
-import { isCustomSides } from './guards'
-
 const generateStandardSides = (sides: number): number[] =>
   Array.from({ length: Number(sides) }, (_, index) => index + 1)
 
@@ -9,8 +6,8 @@ export default abstract class SingleDie<D extends string | number> {
 
   faces: D[]
 
-  constructor(sides: D extends number ? number : CustomSides) {
-    const isCustom = isCustomSides(sides)
+  constructor(sides: D extends number ? number : (number | string)[]) {
+    const isCustom = Array.isArray(sides)
     this.sides = isCustom ? sides.length : Number(sides)
     this.faces = (isCustom ? sides : generateStandardSides(sides)) as D[]
   }
