+++ src/index.ts
@@ -1,6 +1,3 @@
 export * from './Die'
 export { default as roll } from './roll'
-export * from './types/options'
-export * from './types/parameters'
-export * from './types/primitives'
-export * from './types/results'
+export * from './types'
