+++ src/roll/index.ts
@@ -1,5 +1,4 @@
-import { RollArgument } from '../types/argument'
-import { RollResult } from '../types/results'
+import { RollArgument, RollResult } from '../types'
 import generateResult from './generate-result'
 import parseArguments from './parse-arguments'
